Improve client/server code boundary
* should be better security wise
