+++ src/shared/modules/events/PlayersEvent.model.json
@@ -1,6 +1,6 @@
 {
   "ClassName": "RemoteEvent",
   "Properties": {
-    "Name": "PlayersEvent"
+    "Name": "PlayersRE"
   }
 }
