--- src/shared/modules/events/index.ts
@@ -1,7 +1,0 @@
-import { setupPlayerDataEvent } from "./PlayerDataEvent";
-import { setupPlayersEvent } from "./PlayersEvent";
-
-export default function setupEvents() {
-  setupPlayersEvent();
-  setupPlayerDataEvent();
-}
